+++ realm/realm-library/src/androidTest/java/io/realm/IOSRealmTests.java
@@ -55,7 +55,7 @@ public class IOSRealmTests {
     @Rule
     public final TestRealmConfigurationFactory configFactory = new TestRealmConfigurationFactory();
 
-    private static final String[] IOS_VERSIONS = new String[] {"0.98.0"};
+    private static final String[] IOS_VERSIONS = new String[] {"6.0.0-beta.2"};
     private static final String REALM_NAME = "alltypes.realm";
     private Realm realm;
     private Context context;
@@ -162,7 +162,7 @@ public class IOSRealmTests {
             assertFalse(obj.isBoolCol());
             assertEquals(Short.MIN_VALUE, obj.getShortCol());
             assertEquals(Integer.MIN_VALUE, obj.getIntCol());
-            assertEquals(Integer.MIN_VALUE, obj.getLongCol());
+            assertEquals(Long.MIN_VALUE, obj.getLongCol());
             assertEquals(Long.MIN_VALUE, obj.getLongLongCol());
             assertEquals(-Float.MAX_VALUE, obj.getFloatCol(), 0F);
             assertEquals(-Double.MAX_VALUE, obj.getDoubleCol(), 0D);
@@ -173,7 +173,6 @@ public class IOSRealmTests {
     }
 
     @Test
-    @SuppressWarnings("ConstantOverflow")
     public void iOSDataTypesMaximumValues() throws IOException {
         for (String iosVersion : IOS_VERSIONS) {
             configFactory.copyRealmFromAssets(context,
@@ -183,13 +182,13 @@ public class IOSRealmTests {
             IOSAllTypes obj = realm.where(IOSAllTypes.class).findFirst();
             assertEquals(Short.MAX_VALUE, obj.getShortCol());
             assertEquals(Integer.MAX_VALUE, obj.getIntCol());
-            assertEquals(Integer.MAX_VALUE, obj.getLongCol());
+            assertEquals(Long.MAX_VALUE, obj.getLongCol());
             assertEquals(Long.MAX_VALUE, obj.getLongLongCol());
             assertEquals(Float.MAX_VALUE, obj.getFloatCol(), 0F);
             assertEquals(Double.MAX_VALUE, obj.getDoubleCol(), 0D);
             assertArrayEquals(new byte[0], obj.getByteCol());
             assertEquals("", obj.getStringCol());
-            assertEquals(Long.MIN_VALUE, obj.getDateCol().getTime());
+            assertEquals(Long.MAX_VALUE, obj.getDateCol().getTime());
         }
     }