+++ realm/realm-library/src/main/cpp/io_realm_internal_objectstore_OsObjectBuilder.cpp
@@ -158,25 +158,22 @@ JNIEXPORT jlong JNICALL Java_io_realm_internal_objectstore_OsObjectBuilder_nativ
 {
     try {
         SharedRealm shared_realm = *(reinterpret_cast<SharedRealm*>(shared_realm_ptr));
+
+        CreatePolicy policy = CreatePolicy::ForceCreate;
+        if (update_existing && ignore_same_values) {
+            policy = CreatePolicy::UpdateModified;
+        } else if (update_existing) {
+            policy = CreatePolicy::UpdateAll;
+        }
+
         TableRef table = TBL_REF(table_ref_ptr);
         const auto& schema = shared_realm->schema();
         const ObjectSchema& object_schema = get_schema(schema, table);
         JavaContext ctx(env, shared_realm, object_schema);
         auto list = *reinterpret_cast<OsObjectData*>(builder_ptr);
         JavaValue values = JavaValue(list);
-        CreatePolicy policy;
-        if (ignore_same_values) {
-            policy = CreatePolicy::UpdateModified;
-        }
-        else if (update_existing) {
-            policy = CreatePolicy::UpdateAll;
-        }
-        else {
-            policy = CreatePolicy::ForceCreate;
-        }
-
         Object obj = Object::create(ctx, shared_realm, object_schema, values, policy);
-        return reinterpret_cast<jlong>(new Row(obj.row()));
+        return reinterpret_cast<jlong>(new Obj(obj.obj()));
     }
     CATCH_STD()
     return realm::npos;