+++ realm/realm-library/src/main/java/io/realm/internal/ManageableObject.java
@@ -19,7 +19,7 @@ package io.realm.internal;
  * This internal interface represents a java object that corresponds to data
  * that may be managed in the Realm core. It specifies the operations common to all such objects.
  */
-public interface ManagableObject {
+public interface ManageableObject {
 
     /**
      * Checks to see if this object is managed by Realm..