+++ realm/realm-library/src/objectServer/java/io/realm/SyncConfiguration.java
@@ -34,7 +34,6 @@ import java.util.regex.Pattern;
 
 import javax.annotation.Nullable;
 
-import io.realm.annotations.Beta;
 import io.realm.annotations.RealmModule;
 import io.realm.exceptions.RealmException;
 import io.realm.internal.OsRealmConfig;
@@ -231,49 +230,6 @@ public class SyncConfiguration extends RealmConfiguration {
         } catch (URISyntaxException e) {
             throw new IllegalArgumentException("Could not replace '/~/' with a valid user ID.", e);
         }
-    }
-
-    /**
-     * Creates an automatic default configuration based on the the currently logged in user.
-     * <p>
-     * This configuration will point to the default Realm on the server where the user was
-     * authenticated.
-     *
-     * @throws IllegalStateException if no user are logged in, or multiple users have. Only one should
-     * be logged in when calling this method.
-     * @return The constructed {@link SyncConfiguration}.
-     * @deprecated use {@link SyncUser#getDefaultConfiguration()} instead.
-     */
-    @Deprecated
-    @Beta
-    public static SyncConfiguration automatic() {
-        SyncUser user = SyncUser.current();
-        if (user == null) {
-            throw new IllegalStateException("No user was logged in.");
-        }
-        return user.getDefaultConfiguration();
-    }
-
-    /**
-     * Creates an automatic default configuration for the provided user.
-     * <p>
-     * This configuration will point to the default Realm on the server where the user was
-     * authenticated.
-     *
-     * @throws IllegalArgumentException if no user was provided or the user isn't valid.
-     * @return The constructed {@link SyncConfiguration}.
-     * @deprecated use {@link SyncUser#getDefaultConfiguration()} instead.
-     */
-    @Deprecated
-    @Beta
-    public static SyncConfiguration automatic(SyncUser user) {
-        if (user == null) {
-            throw new IllegalArgumentException("Non-null 'user' required.");
-        }
-        if (!user.isValid()) {
-            throw new IllegalArgumentException("User is no logger valid.  Log the user in again.");
-        }
-        return user.getDefaultConfiguration();
     }
 
     // Extract the full server path, minus the file name
@@ -473,20 +429,6 @@ public class SyncConfiguration extends RealmConfiguration {
     }
 
     /**
-     * Whether this configuration is for a query-based Realm.
-     * <p>
-     * Query-based synchronization allows a synchronized Realm to be opened in such a way that
-     * only objects queried by the user are synchronized to the device.
-     *
-     * @return {@code true} to open a query-based Realm {@code false} otherwise.
-     * @deprecated use {@link #isFullySynchronizedRealm()} instead.
-     */
-    @Deprecated
-    public boolean isPartialRealm() {
-        return isPartial;
-    }
-
-    /**
      * Returns whether this configuration is for a fully synchronized Realm or not.
      *
      * @see Builder#fullSynchronization() for more details.
@@ -552,41 +494,6 @@ public class SyncConfiguration extends RealmConfiguration {
         @Nullable // null means the user hasn't explicitly set one. An appropriate default is chosen when calling build()
         private ClientResyncMode clientResyncMode = null;
         private long maxNumberOfActiveVersions = Long.MAX_VALUE;
-
-        /**
-         * Creates an instance of the Builder for the SyncConfiguration. This SyncConfiguration
-         * will be for a fully synchronized Realm.
-         * <p>
-         * Opening a synchronized Realm requires a valid user and an unique URI that identifies that Realm. In URIs,
-         * {@code /~/} can be used as a placeholder for a user ID in case the Realm should only be available to one
-         * user e.g., {@code "realm://objectserver.realm.io/~/default"}.
-         * <p>
-         * The URL cannot end with {@code .realm}, {@code .realm.lock} or {@code .realm.management}.
-         * <p>
-         * The {@code /~/} will automatically be replaced with the user ID when creating the {@link SyncConfiguration}.
-         * <p>
-         * Moreover, the URI defines the local location on disk. The default location of a synchronized Realm file is
-         * {@code /data/data/<packageName>/files/realm-object-server/<user-id>/<last-path-segment>}, but this behavior
-         * can be overwritten using {@link #name(String)} and {@link #directory(File)}.
-         * <p>
-         * Many Android devices are using FAT32 file systems. FAT32 file systems have a limitation that
-         * file names cannot be longer than 255 characters. Moreover, the entire URI should not exceed 256 characters.
-         * If file name and underlying path are too long to handle for FAT32, a shorter unique name will be generated.
-         * See also @{link https://msdn.microsoft.com/en-us/library/aa365247(VS.85).aspx}.
-         *
-         * @param user the user for this Realm. An authenticated {@link SyncUser} is required to open any Realm managed
-         *             by a Realm Object Server.
-         * @param uri URI identifying the Realm. If only a path like {@code /~/default} is given, the configuration will
-         *            assume the file is located on the same server returned by {@link SyncUser#getAuthenticationUrl()}.
-         *
-         * @see SyncUser#isValid()
-         * @deprecated Use {@link SyncUser#createConfiguration(String)} instead.
-         */
-        @Deprecated
-        public Builder(SyncUser user, String uri) {
-            this(BaseRealm.applicationContext, user, uri);
-            fullSynchronization();
-        }
 
         Builder(Context context, SyncUser user, String url) {
             //noinspection ConstantConditions
@@ -1049,18 +956,6 @@ public class SyncConfiguration extends RealmConfiguration {
         }
 
         /**
-         * Setting this will open a query-based Realm.
-         *
-         * @see #isPartialRealm()
-         * @deprecated Use {@link SyncUser#createConfiguration(String)} instead.
-         */
-        @Deprecated
-        public SyncConfiguration.Builder partialRealm() {
-            this.isPartial = true;
-            return this;
-        }
-
-        /**
          * Define this Realm as a fully synchronized Realm.
          * <p>
          * Full synchronization, unlike the default query-based synchronization, will transparently
@@ -1159,6 +1054,28 @@ public class SyncConfiguration extends RealmConfiguration {
                 throw new IllegalArgumentException("Non-null 'mode' required.");
             }
             clientResyncMode = mode;
+            return this;
+        }
+
+        /**
+         * Sets the maximum number of live versions in the Realm file before an {@link IllegalStateException} is thrown when
+         * attempting to write more data.
+         * <p>
+         * Realm is capable of concurrently handling many different versions of Realm objects. This can happen if you
+         * have a Realm open on many different threads or are freezing objects while data is being written to the file.
+         * <p>
+         * Under normal circumstances this is not a problem, but if the number of active versions grow too large, it will
+         * have a negative effect on the filesize on disk. Setting this parameters can therefore be used to prevent uses of
+         * Realm that can result in very large Realms.
+         * <p>
+         * Note, the version number will also increase when changes from other devices are integrated on this device,
+         * so the number of active versions will also depend on what other devices writing to the same Realm are doing.
+         *
+         * @param number the maximum number of active versions before an exception is thrown.
+         * @see <a href="https://realm.io/docs/java/latest/#faq-large-realm-file-size">FAQ</a>
+         */
+        public Builder maxNumberOfActiveVersions(long number) {
+            this.maxNumberOfActiveVersions = number;
             return this;
         }